Finish agent_surfaces: no complex functions, no oversized files
The module started this PR with 48 functions over the complexity threshold, nine
files over 600 lines, and the worst function in the repo. It now has none of any
of them -- `agent_surfaces` drops out of `complex_functions` entirely, and off
the oversized list with it.

This commit is the last seven functions and the four files that decomposing them
pushed over 600:

  event_receiver_service.run           24 -> 8    poll / drain / dispatch / recover
  _transcribe_voice_attachments        21 -> 7    provider, transcribe, record, join
  try_handle_lifecycle                 20 -> 9    one prompt, two reasons for it
  _handle_waiting_event                20 -> 7    narration and prompt are separable
  identity_resolution.resolve          18 -> 7    `_KnownSender` read once, used twice
  handle_platform_webhook              18 -> 10   per-platform special cases named
  send_approval_prompt_for_conversation 17 -> 5   guards, then plan, then deliver
  _collect_new_emails                  16 -> 9    one page's scan is its own function

  progress_observer.py       605 -> 501  + progress_waiting.py (179)
  surface_configuration.py   612 -> 419  + surface_lifecycle.py (228)
  webhook_controller.py      602 -> ~350 + webhook_ingest.py
  event_receiver_service.py  692 -> ~490 + telegram_polling_runner.py

`_publish_native_receiver_event` moved down into `native_receiver_base.py`, the
leaf that exists precisely so a per-platform runner can publish without
importing the coordinator that imports the runners.

Two things worth saying plainly.

The broad-catch *count* went 106 -> 108 without a single new `except Exception`:
the total stayed at 124 the whole time, and splitting a function that held two
of them into two functions holding one each is what the metric saw. Rather than
raise the baseline for that, two genuinely wrong catches are now narrow --
`response.json()` in the Telegram and WhatsApp clients raises
`json.JSONDecodeError`, a `ValueError`, and catching everything there was hiding
our own bugs behind their bad JSON. Total is 122 now.

And the e2e suite caught the split, twice over. Nineteen tests went red on a 405:
the replacement text for `handle_platform_webhook` did not carry its
`@router.post` decorator, so the route silently stopped existing while every
unit test stayed green. `ast` reports a decorated function's lineno at the `def`,
which is the same trap that dropped a `@dataclass` in the Teams commit; the
extraction helper accounts for it on the delete side, and this was the write
side. The whole diff has been audited for other lost decorators -- there are
none.

5,329 unit tests, 173 surface e2e tests, and `make typecheck-critical` pass.
