+++ lemma-backend/architecture-baseline.json
@@ -5,7 +5,7 @@
     "agent:total": 81,
     "agent_surfaces:count": 106,
     "agent_surfaces:max": 5,
-    "agent_surfaces:total": 124,
+    "agent_surfaces:total": 122,
     "apps:count": 4,
     "apps:max": 1,
     "apps:total": 4,
@@ -44,9 +44,6 @@
     "agent:count": 15,
     "agent:max": 42,
     "agent:total": 371,
-    "agent_surfaces:count": 8,
-    "agent_surfaces:max": 24,
-    "agent_surfaces:total": 154,
     "connectors:count": 8,
     "connectors:max": 24,
     "connectors:total": 152,
@@ -95,7 +92,6 @@
     "app/modules/agent_surfaces/api/controllers/surface_controller.py": 627,
     "app/modules/agent_surfaces/api/schemas.py": 610,
     "app/modules/agent_surfaces/platforms/whatsapp/service.py": 623,
-    "app/modules/agent_surfaces/services/event_receiver_service.py": 669,
     "app/modules/agent_surfaces/services/surface_service.py": 1025,
     "app/modules/apps/services/app_service.py": 628,
     "app/modules/connectors/services/connector_operation_service.py": 665,
